rag_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.11.2
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-postgres==0.0.12
openai>=1.10.0
psycopg[binary]==3.1.18
pgvector>=0.2.5,<0.3.0
sqlalchemy==2.0.23
urllib3<2
protobuf>=3.20.3,<6.0.0
python-dotenv==1.0.0